+++ docs/src/changes.md
@@ -1,5 +1,12 @@
 # Changes {#changes}
 
+## v1.9.0
+
+| Date       | Description                                                             |
+|------------|-------------------------------------------------------------------------|
+| 2025-12-11 | Add support for Linux ARM for Python wheels                             |
+| 2025-12-11 | Speed up linux build by using only binary packages for numpy and pandas |
+
 ## v1.8.0
 
 | Date       | Description                 |